feat: make wasm work
