+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "deno.enable": true,
+    "deno.lint": false,
+    "deno.unstable": false
+}
