--- Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "deno-swc"
-version = "0.0.5"
-authors = ["Divy Srivastava <dj.srivastava23@gmail.com>"]
-edition = "2018"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-deno_core = "0.49.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-futures = "0.3.4"
-swc = { git = "https://github.com/swc-project/swc", rev = "053f81c" }
-spack = { git = "https://github.com/swc-project/swc", rev = "053f81c" }
-swc_ecma_visit = "0.8.0"
-swc_ecma_ast = "=0.23.1"
-swc_common = "0.6.2"
-swc_ecma_parser = "0.27.0"
-core = { path = "./swc_core" }
