--- swc_core/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "core"
-version = "0.0.1"
-authors = ["Divy Srivastava <dj.srivastava23@gmail.com>"]
-edition = "2018"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-swc = { git = "https://github.com/swc-project/swc", rev = "053f81c" }
-spack = { git = "https://github.com/swc-project/swc", rev = "053f81c" }
-swc_ecma_visit = "0.8.0"
-swc_ecma_ast = "=0.23.1"
-swc_common = "0.6.2"
-swc_ecma_parser = "0.27.0"
-anyhow = "1.0.31"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-fxhash = "0.2"
-wasm-bindgen = { version = "0.2.64", features = ["serde-serialize"] }
-wee_alloc = { version = "0.4.5", optional = true }
-js-sys = "0.3.44"
-console_error_panic_hook = "0.1.6"
-
-[profile.release]
-lto = true
-opt-level = "z"
-
-[features]
-default = ["wee_alloc"]
