--- swc_deno/print.ts
@@ -1,10 +1,0 @@
-import {
-  swc_print,
-} from "../swc_plugin/index.ts";
-
-export function print(
-  opt: object,
-): { code: string } {
-  const result = JSON.parse(swc_print(opt));
-  return result;
-}
