--- swc_deno/transform.ts
@@ -1,10 +1,0 @@
-import {
-  swc_transform,
-} from "../swc_plugin/index.ts";
-
-export function transform(
-  src: string,
-) {
-  const result = JSON.parse(swc_transform(src));
-  return result;
-}
