+++ swc_wasm/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "deno_swc"
+version = "0.0.1"
+authors = ["Divy Srivastava <dj.srivastava23@gmail.com>"]
+edition = "2018"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+path = "lib.rs"
+
+[dependencies]
+swc = { git = "https://github.com/swc-project/swc", rev = "6b03c65" }
+swc_ecma_ast = { git = "https://github.com/swc-project/swc", rev = "6b03c65" }
+swc_common = { git = "https://github.com/swc-project/swc", rev = "6b03c65" }
+anyhow = "1.0.31"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+fxhash = "0.2"
+wasm-bindgen = { version = "0.2.64", features = ["serde-serialize"] }
+wee_alloc = { version = "0.4.5", optional = true }
+js-sys = "0.3.44"
+console_error_panic_hook = "0.1.6"
+sourcemap = "6.0.1"
+once_cell = "1.3.1"
+
+[profile.release]
+lto = true
+opt-level = "z"
+
+[features]
+default = ["wee_alloc"]
