--- tests/test_files/bundle_me.ts
@@ -1,1 +1,0 @@
-export const z = 3;
